requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-chroma==0.1.4

chromadb==0.5.20
sentence-transformers==3.2.1
rank-bm25==0.2.2
python-dotenv==1.0.1
python-docx==1.1.0
pypdf==4.3.1
redis==5.1.1
groq==0.13.0
openai

